pyproject.toml
[project]
name = "openapi-mcp-generator"
version = "0.1.0"
description = "A tool to generate MCP (Model Context Protocol) servers from OpenAPI specifications"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "jsonref>=1.1.0",
    "mcp>=1.0.0",
    "httpx>=0.24.0",
    "fastmcp>=2.10.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
openapi-to-mcp = "openapi_mcp_generator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
